BUG: fixed issue in generate_output_file(), if path_in and path_out are the same
Former-commit-id: 3c2b374c5aa70268bf773e92c5f1cfbf55bb46f9 [formerly d0fc8d3e187a482b4681d81cdff99fd9dbfeab29]
Former-commit-id: c42d6bee19cca212ebe149c102f650e5654b1753
Former-commit-id: 4198d7029646da47e4edf746b91657388093131f

+++ scripts/sct_utils.py
@@ -121,7 +121,7 @@ def generate_output_file(fname_in, path_out, file_out, ext_out):
     # extract input file extension
     path_in, file_in, ext_in = extract_fname(fname_in)
     # if (i) output path is not local and (ii) output file already exists in nii or nii.gz format, delete it and move file
-    if not path_out == '':
+    if not path_out == '' and not path_in == path_out:
         if os.path.isfile(path_out+file_out+'.nii'):
             os.system('rm '+path_out+file_out+'.nii')
         if os.path.isfile(path_out+file_out+'.nii.gz'):
